fix(core): move fns between files, create command
metriport/metriport-internal#1040

Ref: ENG-257
